overflows
Originally committed as revision 3935 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavformat/asf-enc.c
@@ -700,10 +700,10 @@ static int asf_write_packet(AVFormatContext *s, AVPacket *pkt)
 
     //XXX /FIXME use duration from AVPacket
     if (codec->codec_type == CODEC_TYPE_AUDIO) {
-        duration = (codec->frame_number * codec->frame_size * int64_t_C(10000000)) /
+        duration = (codec->frame_number * (int64_t)codec->frame_size * int64_t_C(10000000)) /
             codec->sample_rate;
     } else {
-        duration = av_rescale(codec->frame_number * codec->frame_rate_base, 10000000, codec->frame_rate);
+        duration = av_rescale(codec->frame_number * (int64_t)codec->frame_rate_base, 10000000, codec->frame_rate);
     }
     if (duration > asf->duration)
         asf->duration = duration;
